Sepearate js file for dynamicHelpText
